+++ pinery-miso/src/main/resources/ca/on/oicr/pinery/lims/miso/queryAllSamples.sql
@@ -70,6 +70,9 @@ SELECT s.alias NAME
         ,COALESCE(pieceRefSlide.name, ssRefSlide.name) reference_slide_id
         ,sssc.targetCellRecovery target_cell_recovery
         ,sssc.cellViability cell_viability
+        ,NULL spike_in
+        ,NULL spike_in_dilution_factor
+        ,NULL spike_in_volume_ul
 FROM Sample s
 LEFT JOIN DetailedSample sai ON sai.sampleId = s.sampleId 
 LEFT JOIN DetailedQcStatus qpd ON qpd.detailedQcStatusId = sai.detailedQcStatusId 
@@ -260,10 +263,14 @@ SELECT l.alias NAME
         ,NULL reference_slide_id
         ,NULL target_cell_recovery
         ,NULL cell_viability
+        ,lsi.alias spike_in
+        ,l.spikeInDilutionFactor spike_in_dilution_factor
+        ,l.spikeInVolume spike_in_volume_ul
 FROM Library l 
 LEFT JOIN Sample parent ON parent.sampleId = l.sample_sampleId
 LEFT JOIN Project sp ON sp.projectId = parent.project_projectId
 LEFT JOIN DetailedLibrary lai ON lai.libraryId = l.libraryId
+LEFT JOIN LibrarySpikeIn lsi ON lsi.spikeInId = l.spikeInId
 LEFT JOIN KitDescriptor kd ON kd.kitDescriptorId = l.kitDescriptorId
 LEFT JOIN LibraryDesignCode ldc ON lai.libraryDesignCodeId = ldc.libraryDesignCodeId
 LEFT JOIN LibraryType lt ON lt.libraryTypeId = l.libraryType
@@ -403,6 +410,9 @@ SELECT d.alias name
         ,NULL reference_slide_id
         ,NULL target_cell_recovery
         ,NULL cell_viability
+        ,NULL spike_in
+        ,NULL spike_in_dilution_factor
+        ,NULL spike_in_volume_ul
 FROM LibraryAliquot d 
 LEFT JOIN LibraryAliquot laParent ON laParent.aliquotId = d.parentAliquotId
 JOIN Library lib ON lib.libraryId = d.libraryId